requirements.txt
# Google API関連
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.4

# Slack SDK
slack_sdk==3.24.0

# 環境変数管理
python-dotenv==1.0.0

# 生成AI関連
google-generativeai==0.3.2

# ドキュメント処理
python-docx==0.8.11

# その他のユーティリティ
requests==2.31.0
urllib3==2.0.7

# オプション: OpenAI APIを使用する場合
# openai==1.3.7
